+++ packages/ui/src/modals/TransferModal/TransactionFailureModal.tsx
@@ -3,26 +3,15 @@ import styled from 'styled-components'
 import { ButtonPrimaryMedium } from '../../components/buttons/Buttons'
 import { Modal, ModalBody, ModalFooter, ModalHeader } from '../../components/modal'
 import { Text } from '../../components/page/Typography/Text'
-import { useAccounts } from '../../hooks/useAccounts'
 
-export function TransactionFailureModal() {
-  const accounts = useAccounts()
+interface Props {
+  onClose: () => void
+}
 
-  if (!accounts.hasAccounts) {
-    return <></>
-  }
-
-  // const [from, to] = accounts.allAccounts
-
+export function TransactionFailureModal({ onClose }: Props) {
   return (
     <Modal>
-      <ModalHeader
-        onClick={() => {
-          /**/
-        }}
-        title="Recovering failure"
-        icon={'😡'}
-      />
+      <ModalHeader onClick={onClose} title="Recovering failure" icon={'😡'} />
       <ModalFailureBody>
         <Text>You have not transferred 'VALUE' balance from 'From acount name' to 'To account name'</Text>
       </ModalFailureBody>
