+++ packages/ui/src/modals/TransferModal/TransferModal.tsx
@@ -2,14 +2,14 @@ import { ISubmittableResult } from '@polkadot/types/types'
 import BN from 'bn.js'
 import React, { useEffect, useState } from 'react'
 import { Observable, Subscription } from 'rxjs'
-import styled from 'styled-components'
-import { BorderRad, Colors } from '../../constants'
 import { Account } from '../../hooks/types'
 import { useKeyring } from '../../hooks/useKeyring'
-import { WaitModal } from '../WaitModal'
 import { SignTransferModal } from './SignTransferModal'
+import { TransactionFailureModal } from './TransactionFailureModal'
+import { TransactionSuccessModal } from './TransactionSuccessModal'
 import { TransferDetailsModal } from './TransferDetailsModal'
 import { WaitForTheExtensionModal } from './WaitForTheExtensionModal'
+import { WaitModal } from '../WaitModal'
 
 interface Props {
   onClose: () => void
@@ -17,7 +17,7 @@ interface Props {
   to?: Account
 }
 
-type ModalState = 'SEND_TOKENS' | 'SIGN_TRANSACTION' | 'EXTENSION_SIGN' | 'SENDING'
+type ModalState = 'SEND_TOKENS' | 'SIGN_TRANSACTION' | 'EXTENSION_SIGN' | 'SENDING' | 'SUCCESS' | 'ERROR'
 
 export function TransferModal({ from, to, onClose }: Props) {
   const keyring = useKeyring()
@@ -54,7 +54,7 @@ export function TransferModal({ from, to, onClose }: Props) {
 
       console.log(`In Block. Block hash: ${status.asInBlock.toString()}`)
 
-      onClose()
+      setStep('SUCCESS')
     }
 
     setSubscription(transaction.subscribe(statusCallback))
@@ -78,101 +78,13 @@ export function TransferModal({ from, to, onClose }: Props) {
     return <WaitForTheExtensionModal />
   }
 
-  return <WaitModal title="Wait for the transaction" description="..." />
+  if (step === 'SENDING') {
+    return <WaitModal title="Wait for the transaction" description="..." />
+  }
+
+  if (step === 'SUCCESS') {
+    return <TransactionSuccessModal onClose={onClose} from={from} to={transferTo} />
+  }
+
+  return <TransactionFailureModal onClose={onClose} />
 }
-
-export const FormLabel = styled.div`
-  font-size: 14px;
-  line-height: 20px;
-  margin-bottom: 8px;
-  font-weight: 700;
-`
-export const Row = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: auto;
-`
-export const AccountRow = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-template-rows: 1fr;
-  align-items: center;
-  min-height: 94px;
-  padding: 16px 132px 16px 14px;
-  border: 1px solid ${Colors.Black[100]};
-  border-radius: ${BorderRad.s};
-  background-color: ${Colors.White};
-`
-export const LockedAccount = styled(AccountRow)`
-  background-color: ${Colors.Black[50]};
-`
-export const TransactionAmount = styled.div`
-  display: grid;
-  grid-template-columns: 284px auto;
-  grid-template-rows: 1fr;
-  grid-column-gap: 24px;
-  align-items: end;
-`
-export const AmountInputBlock = styled.div`
-  display: flex;
-  flex-direction: column;
-`
-
-export const TransactionInfo = styled.div`
-  display: flex;
-  flex-direction: column;
-`
-export const BalanceInfo = styled.div`
-  display: grid;
-  position: relative;
-  grid-template-columns: 1fr 128px;
-  grid-template-rows: 1fr;
-  align-items: center;
-
-  & + & {
-    margin-top: 4px;
-  }
-`
-export const InfoTitle = styled.span`
-  font-size: 10px;
-  line-height: 16px;
-  font-weight: 700;
-  text-transform: uppercase;
-  text-align: right;
-  color: ${Colors.Black[400]};
-`
-export const InfoValue = styled.span`
-  display: grid;
-  position: relative;
-  text-align: right;
-  line-height: 20px;
-`
-
-export const TransactionAmountInfo = styled.div`
-  display: grid;
-  grid-auto-flow: column;
-  grid-column-gap: 12px;
-  width: fit-content;
-  justify-self: center;
-  align-items: center;
-  color: ${Colors.Black[900]};
-`
-
-export const TransactionAmountInfoText = styled.span`
-  padding: 0 8px;
-  font-size: 12px;
-  line-height: 16px;
-  font-weight: 700;
-  color: ${Colors.Black[900]};
-  text-transform: uppercase;
-  border-radius: ${BorderRad.m};
-  background-color: ${Colors.Black[75]};
-
-  .TokenValue {
-    font-size: 12px;
-    line-height: 16px;
-    font-weight: 700;
-    color: ${Colors.Black[900]};
-  }
-`
